add soft_delete mode for date.del, add test, fix existing tests, bump version

+++ src/index.js
@@ -234,14 +234,25 @@ Dare.prototype.del = function del(table, filter, opts = {}) {
 		// Prepare post
 		let a = prepare(query);
 
-		// Construct a db update
-		return this.sql(
+		let sql;
+		if (!opts.soft_delete) {
+			sql =
+				`DELETE FROM ${table}
+				 WHERE
+				 ${serialize(query, '=', 'AND')}
+				 LIMIT ${opts.limit}`;
+		}
+		else {
+			sql =
+				`UPDATE ${table}
+				 SET is_deleted = 1
+				 WHERE
+				 ${serialize(query, '=', 'AND')}
+				 LIMIT ${opts.limit}`;
+		}
 
-			`DELETE FROM ${table}
-			WHERE
-			${serialize(query, '=', 'AND')}
-			LIMIT ${opts.limit}`,
-		a)
+		// Construct a db update
+		return this.sql(sql, a)
 		.then(mustAffectRows);
 	});
 };
